[TASK] Migrate all `ConfigurationRegistry` tests to functional
This class will soon use DI. For that, we'll need to use
functional tests exclusively for testing this class.
