--- Tests/Unit/Configuration/ConfigurationRegistryTest.php
@@ -1,161 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace OliverKlee\Oelib\Tests\Unit\Configuration;
-
-use OliverKlee\Oelib\Configuration\ConfigurationRegistry;
-use OliverKlee\Oelib\Configuration\DummyConfiguration;
-use OliverKlee\Oelib\Configuration\TypoScriptConfiguration;
-use TYPO3\TestingFramework\Core\Unit\UnitTestCase;
-
-/**
- * @covers \OliverKlee\Oelib\Configuration\ConfigurationRegistry
- */
-final class ConfigurationRegistryTest extends UnitTestCase
-{
-    private ConfigurationRegistry $subject;
-
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        $this->subject = new ConfigurationRegistry();
-    }
-
-    protected function tearDown(): void
-    {
-        ConfigurationRegistry::purgeInstance();
-
-        parent::tearDown();
-    }
-
-    /**
-     * @test
-     */
-    public function getInstanceReturnsConfigurationRegistryInstance(): void
-    {
-        self::assertInstanceOf(ConfigurationRegistry::class, ConfigurationRegistry::getInstance());
-    }
-
-    /**
-     * @test
-     */
-    public function getInstanceTwoTimesReturnsSameInstance(): void
-    {
-        self::assertSame(
-            ConfigurationRegistry::getInstance(),
-            ConfigurationRegistry::getInstance(),
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getInstanceAfterPurgeInstanceReturnsNewInstance(): void
-    {
-        $firstInstance = ConfigurationRegistry::getInstance();
-        ConfigurationRegistry::purgeInstance();
-
-        self::assertNotSame(
-            $firstInstance,
-            ConfigurationRegistry::getInstance(),
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getForEmptyNamespaceThrowsException(): void
-    {
-        $this->expectException(\InvalidArgumentException::class);
-        $this->expectExceptionMessage('$namespace must not be empty.');
-        $this->expectExceptionCode(1_331_318_549);
-
-        // @phpstan-ignore-next-line We are explicitly checking for a contract violation here.
-        ConfigurationRegistry::get('');
-    }
-
-    /**
-     * @test
-     */
-    public function setWithEmptyNamespaceThrowsException(): void
-    {
-        $this->expectException(\InvalidArgumentException::class);
-        $this->expectExceptionMessage('$namespace must not be empty.');
-        $this->expectExceptionCode(1_331_318_549);
-
-        // @phpstan-ignore-next-line We are explicitly checking for a contract violation here.
-        ConfigurationRegistry::getInstance()->set('', new DummyConfiguration());
-    }
-
-    /**
-     * @test
-     */
-    public function getAfterSetWithTypoScriptConfigurationReturnsTheSetInstance(): void
-    {
-        $configuration = new TypoScriptConfiguration();
-
-        ConfigurationRegistry::getInstance()->set('foo', $configuration);
-
-        self::assertSame($configuration, ConfigurationRegistry::get('foo'));
-    }
-
-    /**
-     * @test
-     */
-    public function getAfterSetWithDummyConfigurationReturnsTheSetInstance(): void
-    {
-        $configuration = new DummyConfiguration();
-
-        ConfigurationRegistry::getInstance()->set('foo', $configuration);
-
-        self::assertSame($configuration, ConfigurationRegistry::get('foo'));
-    }
-
-    /**
-     * @test
-     */
-    public function setTwoTimesForTheSameNamespaceDoesNotFail(): void
-    {
-        $this->subject->set('foo', new DummyConfiguration());
-        $this->subject->set('foo', new DummyConfiguration());
-    }
-
-    /**
-     * @test
-     */
-    public function getByNamespaceForEmptyNamespaceThrowsException(): void
-    {
-        $this->expectException(\InvalidArgumentException::class);
-        $this->expectExceptionMessage('$namespace must not be empty.');
-        $this->expectExceptionCode(1_331_318_549);
-
-        // @phpstan-ignore-next-line We are explicitly checking for a contract violation here.
-        $this->subject->getByNamespace('');
-    }
-
-    /**
-     * @test
-     */
-    public function getByNamespaceAfterSetWithTypoScriptConfigurationReturnsTheSetInstance(): void
-    {
-        $configuration = new TypoScriptConfiguration();
-
-        $this->subject->set('foo', $configuration);
-
-        self::assertSame($configuration, $this->subject->getByNamespace('foo'));
-    }
-
-    /**
-     * @test
-     */
-    public function getByNamespaceAfterSetWithDummyConfigurationReturnsTheSetInstance(): void
-    {
-        $configuration = new DummyConfiguration();
-
-        $this->subject->set('foo', $configuration);
-
-        self::assertSame($configuration, $this->subject->getByNamespace('foo'));
-    }
-}
